refactor(#20): GotoTopButton - ui, 상태 로직 파일 분리
